+++ alacritty/src/display/content.rs
@@ -18,14 +18,11 @@ use alacritty_terminal::term::{
 use crate::config::ui_config::UiConfig;
 use crate::display::color::{List, DIM_FACTOR};
 use crate::display::hint::HintState;
-use crate::display::Display;
+use crate::display::{self, Display, MAX_SEARCH_LINES};
 use crate::event::SearchState;
 
 /// Minimum contrast between a fixed cursor color and the cell's background.
 pub const MIN_CURSOR_CONTRAST: f64 = 1.5;
-
-/// Maximum number of linewraps followed outside of the viewport during search highlighting.
-const MAX_SEARCH_LINES: usize = 100;
 
 /// Renderable terminal content.
 ///
@@ -105,14 +102,6 @@ impl<'a> RenderableContent<'a> {
             return None;
         }
 
-        // Expand across wide cell when inside wide char or spacer.
-        let is_wide = if cell.flags.contains(Flags::WIDE_CHAR_SPACER) {
-            self.terminal_cursor.point.column -= 1;
-            true
-        } else {
-            cell.flags.contains(Flags::WIDE_CHAR)
-        };
-
         // Cursor colors.
         let color = if self.terminal_content.mode.contains(TermMode::VI) {
             self.config.ui_config.colors.vi_mode_cursor
@@ -138,14 +127,14 @@ impl<'a> RenderableContent<'a> {
 
         // Convert cursor point to viewport position.
         let cursor_point = self.terminal_cursor.point;
-        let line = (cursor_point.line + self.terminal_content.display_offset as i32).0 as usize;
-        let point = Point::new(line, cursor_point.column);
+        let display_offset = self.terminal_content.display_offset;
+        let point = display::point_to_viewport(display_offset, cursor_point).unwrap();
 
         Some(RenderableCursor {
+            is_wide: cell.flags.contains(Flags::WIDE_CHAR),
             shape: self.terminal_cursor.shape,
             cursor_color,
             text_color,
-            is_wide,
             point,
         })
     }
@@ -219,10 +208,12 @@ impl RenderableCell {
             .selection
             .map_or(false, |selection| selection.contains_cell(&cell, content.terminal_cursor));
 
+        let display_offset = content.terminal_content.display_offset;
+        let viewport_start = Point::new(Line(-(display_offset as i32)), Column(0));
+        let colors = &content.config.ui_config.colors;
         let mut character = cell.c;
 
-        let colors = &content.config.ui_config.colors;
-        if let Some((c, is_first)) = content.hint.advance(cell.point) {
+        if let Some((c, is_first)) = content.hint.advance(viewport_start, cell.point) {
             let (config_fg, config_bg) = if is_first {
                 (colors.hints.start.foreground, colors.hints.start.background)
             } else {
@@ -243,23 +234,18 @@ impl RenderableCell {
                 bg_alpha = 1.0;
             }
         } else if content.search.advance(cell.point) {
-            // Highlight the cell if it is part of a search match.
-            let config_fg = colors.search.matches.foreground;
-            let config_bg = colors.search.matches.background;
+            let focused = content.focused_match.map_or(false, |fm| fm.contains(&cell.point));
+            let (config_fg, config_bg) = if focused {
+                (colors.search.focused_match.foreground, colors.search.focused_match.background)
+            } else {
+                (colors.search.matches.foreground, colors.search.matches.background)
+            };
             Self::compute_cell_rgb(&mut fg, &mut bg, &mut bg_alpha, config_fg, config_bg);
-
-            // Apply the focused match colors, using the normal match colors as base reference.
-            if content.focused_match.map_or(false, |fm| fm.contains(&cell.point)) {
-                let config_fg = colors.search.focused_match.foreground;
-                let config_bg = colors.search.focused_match.background;
-                Self::compute_cell_rgb(&mut fg, &mut bg, &mut bg_alpha, config_fg, config_bg);
-            }
         }
 
         // Convert cell point to viewport position.
         let cell_point = cell.point;
-        let line = (cell_point.line + content.terminal_content.display_offset as i32).0 as usize;
-        let point = Point::new(line, cell_point.column);
+        let point = display::point_to_viewport(display_offset, cell_point).unwrap();
 
         RenderableCell {
             zerowidth: cell.zerowidth().map(|zerowidth| zerowidth.to_vec()),
@@ -409,7 +395,7 @@ impl<'a> Hint<'a> {
     /// this position will be returned.
     ///
     /// The tuple's [`bool`] will be `true` when the character is the first for this hint.
-    fn advance(&mut self, point: Point) -> Option<(char, bool)> {
+    fn advance(&mut self, viewport_start: Point, point: Point) -> Option<(char, bool)> {
         // Check if we're within a match at all.
         if !self.regex.advance(point) {
             return None;
@@ -420,7 +406,7 @@ impl<'a> Hint<'a> {
             .regex
             .matches
             .get(self.regex.index)
-            .map(|regex_match| regex_match.start())
+            .map(|regex_match| max(*regex_match.start(), viewport_start))
             .filter(|start| start.line == point.line)?;
 
         // Position within the hint label.
@@ -441,7 +427,7 @@ impl<'a> From<&'a HintState> for Hint<'a> {
 
 /// Wrapper for finding visible regex matches.
 #[derive(Default, Clone)]
-pub struct RegexMatches(Vec<RangeInclusive<Point>>);
+pub struct RegexMatches(pub Vec<RangeInclusive<Point>>);
 
 impl RegexMatches {
     /// Find all visible matches.