+++ CHANGELOG.md
@@ -17,6 +17,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Synchronized terminal updates using `DCS = 1 s ST`/`DCS = 2 s ST`
 - Regex terminal hints ([see features.md](./docs/features.md#hints))
 - macOS keybinding (cmd+alt+H) hiding all windows other than Alacritty
+- Support for `magnet` URLs
+
+### Changed
+
+- The vi mode cursor is now created in the top-left if the terminal cursor is invisible
+- Focused search match will use cell instead of match colors for CellForeground/CellBackground
+- URL highlighting has moved from `mouse.url` to the `hints` config section
 
 ### Fixed
 
@@ -26,6 +33,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Clicking on URLs on Windows incorrectly opens File Explorer
 - Incorrect underline cursor thickness on wide cell
 - Viewport moving around when resizing while scrolled into history
+- Block cursor not expanding across fullwidth characters when on the right side of it
+- Overwriting fullwidth characters only clearing one of the involved cells
 
 ### Removed